requirements.txt
# Django trunk
svn+http://code.djangoproject.com/svn/django/trunk/#egg=Django==1.4c2

# Django stable
# django

# Bootstrap theme and such
django-bootstrapped

# Backbone front-end application framework
git+git://github.com/mjumbewu/django-backbonejs.git#egg=django-backbonejs

# Mustache templates for the front-end
django-mustachejs

# The REST framework to support an API
djangorestframework

# Database admin
south
psycopg2

# GIS
simplejson
vectorformats

# Testing
django-nose
django-jasmine
django-debug-toolbar

# Misc
bpython
